moved editor.vue
