+++ graphery/src/components/framework/Editor.vue
